requirements.txt
pdfplumber==0.11.4
pytesseract==0.3.13
Pillow==11.0.0
python-docx==1.1.2
docxtpl==0.18.0
docx2pdf==0.1.8
opencv-python==4.10.0.84
streamlit==1.41.1
